Added ability to use FastHPOCR instead of exact string matching. Refactored HPOConceptRecognizer inheritance. Added data and tests.
